Refactoring Version advanced to 0.92
