+++ grid-renderers-collection-addon/pom.xml
@@ -5,13 +5,13 @@
 	<groupId>com.vaadin</groupId>
 	<artifactId>grid-renderers-collection-addon</artifactId>
 	<packaging>jar</packaging>
-	<version>0.91</version>
+	<version>0.92</version>
 	<name>Grid Renderers collection For Vaadin7</name>
 
 	<properties>
 		<project.build.sourceEncoding>UTF-8</project.build.sourceEncoding>
-		<vaadin.version>7.5.6</vaadin.version>
-		<vaadin.plugin.version>7.5.6</vaadin.plugin.version>
+		<vaadin.version>7.5.7</vaadin.version>
+		<vaadin.plugin.version>${vaadin.version}</vaadin.plugin.version>
 
 		<!-- ZIP Manifest fields -->
 		<Implementation-Version>${project.version}</Implementation-Version>
