+++ grid-renderers-collection-addon/src/main/java/org/vaadin/grid/cellrenderers/client/editable/DateFieldRendererServerRpc.java
@@ -1,4 +1,4 @@
-package org.vaadin.grid.cellrenderers.client;
+package org.vaadin.grid.cellrenderers.client.editable;
 
 import java.util.Date;
