+++ grid-renderers-collection-addon/src/main/java/org/vaadin/grid/cellrenderers/editoraware/CheckboxRenderer.java
@@ -13,7 +13,7 @@
  * License for the specific language governing permissions and limitations under
  * the License.
  */
-package org.vaadin.grid.cellrenderers;
+package org.vaadin.grid.cellrenderers.editoraware;
 
 import com.vaadin.data.Container;
 import com.vaadin.data.Property;
